Code refactoring on human player module; code mofification for compile warning of computer player module; Added unit test files for computer player module.
