+++ include/return_types.h
@@ -17,12 +17,11 @@ typedef enum {
     RET_SUCCESS,                //0
     RET_FAILURE,                //1 - Generic failure
     RET_BUSY,                   //2
-    RET_TIMEOUT,                //3
-    RET_INVALID_LEN,            //4
-    RET_ARG_NULL,               //5
-    RET_INSUFFICIENT_MEM,       //6
-    RET_INVALID_CARD,           //7
-    RET_EMPTY_LIST,             //8 
+    RET_INVALID_LEN,            //3
+    RET_ARG_NULL,               //4
+    RET_INSUFFICIENT_MEM,       //5
+    RET_INVALID_CARD,           //6
+    RET_EMPTY_LIST,             //7 
 }
 ret_type_e;
