+++ test/include/test_computer_player.h
@@ -1,0 +1,5 @@
+#ifndef __TEST_COMPUTER_PLAYER_HEADER__
+#define __TEST_COMPUTER_PLAYER_HEADER__
+
+int test_computer_player(void);
+#endif // __TEST_COMPUTER_PLAYER_HEADER__
